Original change: https://android-review.googlesource.com/c/platform/system/extras/+/1715287

+++ f2fs_utils/mkf2fsuserimg.sh
@@ -8,7 +8,7 @@ Usage:
 ${0##*/} OUTPUT_FILE SIZE
          [-S] [-C FS_CONFIG] [-f SRC_DIR] [-D PRODUCT_OUT]
          [-s FILE_CONTEXTS] [-t MOUNT_POINT] [-T TIMESTAMP]
-         [-L LABEL] [--prjquota] [--casefold] [--compression]
+         [-L LABEL] [--prjquota] [--casefold] [--compression] [--readonly]
          [--sldc <num> [sload compression sub-options]]
 <num>: number of the sload compression args, e.g.  -a LZ4 counts as 2
        when sload compression args are not given, <num> must be 0,
@@ -95,6 +95,11 @@ if [[ "$1" == "--compression" ]]; then
   MKFS_OPTS+=" -O compression,extra_attr"
   shift;
 fi
+if [[ "$1" == "--readonly" ]]; then
+  MKFS_OPTS+=" -O ro"
+  READONLY=1
+  shift;
+fi
 
 if [[ "$1" == "--sldc" ]]; then
   if [ -z "$COMPRESS_SUPPORT" ]; then
@@ -122,32 +127,54 @@ if [ -z $SIZE ]; then
   exit 2
 fi
 
-if [ "$SPARSE_IMG" = "false" ]; then
+function _truncate()
+{
+  if [ "$SPARSE_IMG" = "true" ]; then
+    return
+  fi
+
   TRUNCATE_CMD="truncate -s $SIZE $OUTPUT_FILE"
   echo $TRUNCATE_CMD
   $TRUNCATE_CMD
   if [ $? -ne 0 ]; then
     exit 3
   fi
-fi
+}
 
-MAKE_F2FS_CMD="make_f2fs -g android $MKFS_OPTS $OUTPUT_FILE"
-echo $MAKE_F2FS_CMD
-$MAKE_F2FS_CMD
-if [ $? -ne 0 ]; then
-  if [ "$SPARSE_IMG" = "false" ]; then
-    rm -f $OUTPUT_FILE
+function _build()
+{
+  MAKE_F2FS_CMD="make_f2fs -g android $MKFS_OPTS $OUTPUT_FILE"
+  echo $MAKE_F2FS_CMD
+  $MAKE_F2FS_CMD
+  if [ $? -ne 0 ]; then
+    if [ "$SPARSE_IMG" = "false" ]; then
+      rm -f $OUTPUT_FILE
+    fi
+    exit 4
   fi
-  exit 4
-fi
 
-SLOAD_F2FS_CMD="sload_f2fs $SLOAD_OPTS $OUTPUT_FILE"
-echo $SLOAD_F2FS_CMD
-$SLOAD_F2FS_CMD
-# allow 1: Filesystem errors corrected
-ret=$?
-if [ $ret -ne 0 ] && [ $ret -ne 1 ]; then
-  rm -f $OUTPUT_FILE
-  exit 4
+  SLOAD_F2FS_CMD="sload_f2fs $SLOAD_OPTS $OUTPUT_FILE"
+  echo $SLOAD_F2FS_CMD
+  MB_SIZE=`$SLOAD_F2FS_CMD | grep "Max image size" | awk '{print $5}'`
+  # allow 1: Filesystem errors corrected
+  ret=$?
+  if [ $ret -ne 0 ] && [ $ret -ne 1 ]; then
+    rm -f $OUTPUT_FILE
+    exit 4
+  fi
+  SIZE=$(((MB_SIZE + 6) * 1024 * 1024))
+}
+
+_truncate
+_build
+
+# readonly + compress can reduce the image
+if [ "$READONLY" ] && [ "$COMPRESS_SUPPORT" ]; then
+  if [ "$SPARSE_IMG" = "true" ]; then
+    MKFS_OPTS+=" -S $SIZE"
+    rm -f $OUTPUT_FILE && touch $OUTPUT_FILE
+  fi
+  _truncate
+  _build
 fi
 exit 0
